gentoo-info: Add support for CPE URIs
To make it easy to detect issues with the installed versions.
Automatically inserting the installed version in the CPE (the Gentoo
metadata is per-project, not per-version) may be possible, but I
haven't worked that out yet.  Also rename the script from
gentoo-licenses -> gentoo-info now that the output is more generic.